Create generate_primes
This currently only works for libgmp... We need to find some way to convert to uberzahl, but Piazza said it should be possible. This runs MUCH faster than the version used currently. Running for largest parameters (L=3072, N=256) maybe takes 5-7 seconds. I've tested it for basic correctness, and it seems to work. This file generates the appropriate p and q. While we're add it, we might as well add finding g to this. I'll look at that tomorrow/Friday.

Also, turns out I don't know how to use Github. All I know how to do is add new files. ha.